requirements.txt
# Sentiment-Aware Real Estate Intelligence System
# Python Dependencies

# ============================================================================
# Data Processing
# ============================================================================
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0
openpyxl>=3.1.0,<4.0.0

# ============================================================================
# NLP & Text Processing
# ============================================================================
nltk>=3.8.0,<4.0.0
vaderSentiment>=3.3.2,<4.0.0
rapidfuzz>=3.0.0,<4.0.0
spacy>=3.5.0,<4.0.0

# ============================================================================
# Machine Learning
# ============================================================================
scikit-learn>=1.3.0,<2.0.0
xgboost>=2.0.0,<3.0.0
joblib>=1.3.0,<2.0.0

# ============================================================================
# AI Integration (Optional - uses VADER fallback if not available)
# ============================================================================
google-generativeai>=0.3.0,<1.0.0

# ============================================================================
# Dashboard & Visualization
# ============================================================================
streamlit>=1.28.0,<2.0.0
plotly>=5.18.0,<6.0.0
pydeck>=0.8.0,<1.0.0

# ============================================================================
# Utilities
# ============================================================================
python-dotenv>=1.0.0,<2.0.0
tqdm>=4.66.0,<5.0.0
requests>=2.31.0,<3.0.0

# ============================================================================
# Development & Testing (Optional)
# ============================================================================
pytest>=7.4.0,<8.0.0
